+++ effectful/internals/disassembly.py
